requirement.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
pandas
joblib
scikit-learn
python-jose[cryptography]
python-multipart